+++ CHANGELOG.md
@@ -246,3 +246,4 @@
 - 2024-01-10: Add mutate, mod, tz country, NOTICE, +4 db
 - 2024-01-11: Impl bettercontact, LICENSE, frontier, find email, +6 query
 - 2024-01-11: Retry configuration, lead, email opener, NOTICE, +6
+- 2024-01-12: Impl descend, mutate, discover, deal, +4 query